# swagger.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/users":{"post":{"operationId":"UsersController_create","summary":"Create user","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"200":{"description":"Create user","content":{"application/json":{"schema":{"type":"string"}}}},"400":{"description":"Bad request"}}},"get":{"operationId":"UsersController_findAll","summary":"Get all users","parameters":[{"name":"page","required":true,"in":"query","schema":{"type":"number"}},{"name":"limit","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"Get all users","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}}}}},"/users/{id}":{"get":{"operationId":"UsersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"put":{"operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}}},"delete":{"operationId":"UsersController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}}},"info":{"title":"Users API","description":"The users API description","version":"1.0","contact":{}},"tags":[{"name":"users","description":""}],"servers":[],"components":{"schemas":{"CreateUserDto":{"type":"object","properties":{"name":{"type":"string","example":"Brandi Walsh"},"email":{"type":"string","example":"Una19@yahoo.com"},"password":{"type":"string","example":"QYeN9XQ2OHubmmw"},"age":{"type":"number","example":30},"city":{"type":"string","example":"New York"}},"required":["name","email","password","age","city"]},"UserResponseDto":{"type":"object","properties":{"name":{"type":"string","example":"Amelia Fisher-Streich","description":"The user name"},"email":{"type":"string","example":"Melissa_Gusikowski@hotmail.com","description":"The user email","uniqueItems":true},"age":{"type":"number","example":36,"description":"The user age"},"city":{"type":"string","example":"Greenland","description":"The user city"},"limit":{"type":"number","example":1,"description":"Maximum number of returned users"},"page":{"type":"number","example":1,"description":"Page number"}}},"UpdateUserDto":{"type":"object","properties":{"name":{"type":"string","example":"Brandi Walsh"},"email":{"type":"string","example":"Una19@yahoo.com"},"password":{"type":"string","example":"QYeN9XQ2OHubmmw"},"age":{"type":"number","example":30},"city":{"type":"string","example":"New York"}}}}}}